Consolidated STK Swift stuff into one file
