--- Sources/AudioKit/Nodes/Generators/Physical Models/Clarinet.swift
@@ -1,17 +1,0 @@
-// Copyright AudioKit. All Rights Reserved. Revision History at http://github.com/AudioKit/AudioKit/
-
-import AVFoundation
-
-/// STK Clarinet
-///
-public class Clarinet: Node, MIDITriggerable {
-
-    /// Connected nodes
-    public var connections: [Node] { [] }
-    
-    /// Underlying AVAudioNode
-    public var avAudioNode = instantiate(instrument: "clar")
-    
-    /// Initialize the STK Clarinet model
-    public init() {}
-}
