--- Sources/AudioKit/Nodes/Generators/Physical Models/MandolinString.swift
@@ -1,17 +1,0 @@
-// Copyright AudioKit. All Rights Reserved. Revision History at http://github.com/AudioKit/AudioKit/
-
-import AVFoundation
-
-/// STK Mandole
-///
-public class MandolinString: Node, MIDITriggerable {
-
-    /// Connected nodes
-    public var connections: [Node] { [] }
-    
-    /// Underlying AVAudioNode
-    public var avAudioNode = instantiate(instrument: "mand")
-    
-    /// Initialize the STK Mandolin model
-    public init() {}
-}
