--- Sources/AudioKit/Nodes/Generators/Physical Models/TubularBells.swift
@@ -1,17 +1,0 @@
-// Copyright AudioKit. All Rights Reserved. Revision History at http://github.com/AudioKit/AudioKit/
-
-import AVFoundation
-
-/// STK Tubuluar Bells
-///
-public class TubularBells: Node, MIDITriggerable {
-    
-    /// Connected nodes
-    public var connections: [Node] { [] }
-    
-    /// Underlying AVAudioNode
-    public var avAudioNode = instantiate(instrument: "tbel")
-
-    /// Initialize the STK Tubular Bells model
-    public init() { }
-}
